# Use mc.buf as scratchpad for intarpolation