```
benchmark                  old ns/op     new ns/op     delta
BenchmarkInterpolation     1900          1403          -26.16%

benchmark                  old allocs     new allocs     delta
BenchmarkInterpolation     2              1              -50.00%

benchmark                  old bytes     new bytes     delta
BenchmarkInterpolation     448           160           -64.29%
```