@methane 2015-02-18
I can remove last one allocation: Stop using `mc.exec()` and build packet directly.
But it may be dirty a little.  May I do it?