@arnehormann 2015-02-18
I'm opposed to it, I like having the packet building in one place.
But I do like the part where you use mc.buf.takeSmallBuffer (or takeCompleteBuffer) as a scratch space... the packet construction code doesn't need it yet.
By the way - why did you change buffer.go?