@methane 2015-02-18
@arnehormann 

>  But I do like the part where you use mc.buf.takeSmallBuffer (or takeCompleteBuffer) as a scratch space... the packet construction code doesn't need it yet.

packet size is len(query) + 4.  This size is used just after interpolate.
So I think reserving same size is good idea.

>  By the way - why did you change buffer.go?

fixed.