@methane 2015-02-18
Now `interpolateParams()` uses `mc.buf.takeCompleteBuffer()`.
And I've eliminated `estimateParamLength()`.